move cheatcodes
