+++ protostar/commands/test/environments/common_test_cheatcode_factory.py
@@ -2,7 +2,11 @@ from typing import List
 
 from starkware.starknet.business_logic.execution.objects import CallInfo
 
-from protostar.commands.test.cheatcodes import (
+from protostar.commands.test.starkware.test_execution_state import TestExecutionState
+from protostar.starknet.cheatcode import Cheatcode
+from protostar.starknet.cheatcode_factory import CheatcodeFactory
+from protostar.starknet.hint_local import HintLocal
+from protostar.test_runner.cheatcodes import (
     DeclareCheatcode,
     DeployCheatcode,
     DeployContractCheatcode,
@@ -15,11 +19,7 @@ from protostar.commands.test.cheatcodes import (
     StoreCheatcode,
     WarpCheatcode,
 )
-from protostar.commands.test.cheatcodes.reflect.cairo_struct import CairoStructHintLocal
-from protostar.commands.test.starkware.test_execution_state import TestExecutionState
-from protostar.starknet.cheatcode import Cheatcode
-from protostar.starknet.cheatcode_factory import CheatcodeFactory
-from protostar.starknet.hint_local import HintLocal
+from protostar.test_runner.cheatcodes.reflect.cairo_struct import CairoStructHintLocal
 from protostar.test_runner.test_context import TestContextHintLocal
