+++ protostar/commands/test/environments/setup_execution_environment.py
@@ -2,13 +2,13 @@ from typing import List
 
 from starkware.starknet.business_logic.execution.objects import CallInfo
 
-from protostar.commands.test.cheatcodes import MaxExamplesCheatcode
 from protostar.commands.test.environments.common_test_cheatcode_factory import (
     CommonTestCheatcodeFactory,
 )
 from protostar.commands.test.starkware.test_execution_state import TestExecutionState
 from protostar.starknet.cheatcode import Cheatcode
 from protostar.starknet.execution_environment import ExecutionEnvironment
+from protostar.test_runner.cheatcodes import MaxExamplesCheatcode
 
 
 class SetupExecutionEnvironment(ExecutionEnvironment[None]):
